Add support for non-visible works
