+++ catalogue_graph/src/ingestor/models/query/work.py
@@ -1,6 +1,8 @@
 from pydantic import Field
 
+from ingestor.extractors.works_extractor import VisibleExtractedWork
 from ingestor.models.shared.serialisable import ElasticsearchModel
+from ingestor.transformers.work_query_transformer import QueryWorkTransformer
 
 
 class QueryWork(ElasticsearchModel):
@@ -42,3 +44,35 @@ class QueryWork(ElasticsearchModel):
         serialization_alias="collectionPath.label"
     )
     collection_path_path: str | None = Field(serialization_alias="collectionPath.path")
+
+    @classmethod
+    def from_extracted_work(cls, extracted: VisibleExtractedWork) -> "QueryWork":
+        work = extracted.work
+        transformer = QueryWorkTransformer(extracted)
+
+        return QueryWork(
+            id=work.state.canonical_id,
+            collection_path_label=transformer.collection_path_label,
+            collection_path_path=transformer.collection_path,
+            alternative_titles=work.data.alternative_titles,
+            contributors_agent_label=transformer.contributor_agent_labels,
+            genres_concepts_label=list(transformer.genre_concept_labels),
+            subjects_concepts_label=list(transformer.subject_concept_labels),
+            description=work.data.description,
+            edition=work.data.edition,
+            source_identifier_value=work.state.source_identifier.value,
+            identifiers_value=list(transformer.identifiers),
+            images_id=transformer.image_ids,
+            images_identifiers_value=list(transformer.image_source_identifiers),
+            items_id=list(transformer.item_ids),
+            items_identifiers_value=list(transformer.item_identifiers),
+            items_shelfmarks_value=list(transformer.item_shelfmarks),
+            languages_label=[i.label for i in work.data.languages],
+            lettering=work.data.lettering,
+            notes_contents=[n.contents for n in work.data.notes],
+            production_label=list(transformer.production_labels),
+            part_of_title=list(transformer.part_of_titles),
+            physical_description=work.data.physical_description,
+            reference_number=work.data.reference_number,
+            title=work.data.title,
+        )
