+++ catalogue_graph/src/ingestor/models/shared/deleted_reason.py
@@ -1,0 +1,8 @@
+from typing import Literal
+
+from pydantic import BaseModel
+
+
+class DeletedReason(BaseModel):
+    type: Literal["DeletedFromSource", "SuppressedFromSource", "TeiDeletedInMerger"]
+    info: str | None = None
