+++ catalogue_graph/src/ingestor/models/shared/invisible_reason.py
@@ -1,0 +1,16 @@
+from typing import Literal
+
+from pydantic import BaseModel
+
+
+class InvisibleReason(BaseModel):
+    type: Literal[
+        "CopyrightNotCleared",
+        "SourceFieldMissing",
+        "InvalidValueInSourceField",
+        "UnlinkedHistoricalLibraryMiro",
+        "UnableToTransform",
+        "MetsWorksAreNotVisible",
+    ]
+    info: str | None = None
+    message: str | None = None
